+++ internal/controller/inferenceservice_deployment_test.go
@@ -17,110 +17,17 @@ limitations under the License.
 package controller
 
 import (
-	"context"
-	"fmt"
-
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	inferencev1alpha1 "github.com/defilantech/llmkube/api/v1alpha1"
 )
 
-var _ = Describe("InferenceService Controller", func() {
-	Context("When reconciling a resource", func() {
-		const resourceName = "test-resource"
-		const modelName = "test-model"
-
-		ctx := context.Background()
-
-		typeNamespacedName := types.NamespacedName{
-			Name:      resourceName,
-			Namespace: "default",
-		}
-		modelNamespacedName := types.NamespacedName{
-			Name:      modelName,
-			Namespace: "default",
-		}
-		inferenceservice := &inferencev1alpha1.InferenceService{}
-
-		BeforeEach(func() {
-			By("creating a Model resource first")
-			model := &inferencev1alpha1.Model{}
-			err := k8sClient.Get(ctx, modelNamespacedName, model)
-			if err != nil && errors.IsNotFound(err) {
-				modelResource := &inferencev1alpha1.Model{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      modelName,
-						Namespace: "default",
-					},
-					Spec: inferencev1alpha1.ModelSpec{
-						Source:       "https://huggingface.co/test/model.gguf",
-						Format:       "gguf",
-						Quantization: "Q4_K_M",
-						Hardware:     &inferencev1alpha1.HardwareSpec{Accelerator: "cpu"},
-						Resources:    &inferencev1alpha1.ResourceRequirements{CPU: "1", Memory: "1Gi"},
-					},
-				}
-				Expect(k8sClient.Create(ctx, modelResource)).To(Succeed())
-			}
-
-			By("creating the custom resource for the Kind InferenceService")
-			err = k8sClient.Get(ctx, typeNamespacedName, inferenceservice)
-			if err != nil && errors.IsNotFound(err) {
-				replicas := int32(1)
-				resource := &inferencev1alpha1.InferenceService{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      resourceName,
-						Namespace: "default",
-					},
-					Spec: inferencev1alpha1.InferenceServiceSpec{
-						ModelRef: modelName,
-						Replicas: &replicas,
-						Image:    "ghcr.io/ggml-org/llama.cpp:server",
-					},
-				}
-				Expect(k8sClient.Create(ctx, resource)).To(Succeed())
-			}
-		})
-
-		AfterEach(func() {
-			By("Cleanup the specific resource instance InferenceService")
-			resource := &inferencev1alpha1.InferenceService{}
-			err := k8sClient.Get(ctx, typeNamespacedName, resource)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(k8sClient.Delete(ctx, resource)).To(Succeed())
-
-			By("Cleanup the Model resource")
-			modelResource := &inferencev1alpha1.Model{}
-			err = k8sClient.Get(ctx, modelNamespacedName, modelResource)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(k8sClient.Delete(ctx, modelResource)).To(Succeed())
-		})
-		It("should successfully reconcile the resource", func() {
-			By("Reconciling the created resource")
-			controllerReconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-
-			_, err := controllerReconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: typeNamespacedName,
-			})
-			Expect(err).NotTo(HaveOccurred())
-			// TODO(user): Add more specific assertions depending on your controller's reconciliation logic.
-			// Example: If you expect a certain status condition after reconciliation, verify it here.
-		})
-	})
-})
 
 var _ = Describe("calculateTensorSplit", func() {
 	Context("with different GPU counts", func() {
@@ -171,6 +78,7 @@ var _ = Describe("calculateTensorSplit", func() {
 		})
 	})
 })
+
 
 var _ = Describe("Multi-GPU Deployment Construction", func() {
 	Context("when constructing a deployment with multi-GPU model", func() {
@@ -637,6 +545,7 @@ var _ = Describe("Multi-GPU Deployment Construction", func() {
 		})
 	})
 })
+
 
 var _ = Describe("Context Size Configuration", func() {
 	Context("when constructing a deployment with context size", func() {
@@ -1373,144 +1282,6 @@ var _ = Describe("Context Size Configuration", func() {
 	})
 })
 
-var _ = Describe("Multi-GPU End-to-End Reconciliation", func() {
-	Context("when reconciling a multi-GPU InferenceService", func() {
-		const multiGPUModelName = "e2e-multi-gpu-model"
-		const multiGPUServiceName = "e2e-multi-gpu-service"
-
-		ctx := context.Background()
-
-		modelNamespacedName := types.NamespacedName{
-			Name:      multiGPUModelName,
-			Namespace: "default",
-		}
-		serviceNamespacedName := types.NamespacedName{
-			Name:      multiGPUServiceName,
-			Namespace: "default",
-		}
-
-		BeforeEach(func() {
-			By("creating a multi-GPU Model resource")
-			model := &inferencev1alpha1.Model{}
-			err := k8sClient.Get(ctx, modelNamespacedName, model)
-			if err != nil && errors.IsNotFound(err) {
-				modelResource := &inferencev1alpha1.Model{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      multiGPUModelName,
-						Namespace: "default",
-					},
-					Spec: inferencev1alpha1.ModelSpec{
-						Source:       "https://huggingface.co/test/multi-gpu-model.gguf",
-						Format:       "gguf",
-						Quantization: "Q4_K_M",
-						Hardware: &inferencev1alpha1.HardwareSpec{
-							Accelerator: "cuda",
-							GPU: &inferencev1alpha1.GPUSpec{
-								Enabled: true,
-								Count:   2,
-								Vendor:  "nvidia",
-								Layers:  -1,
-								Sharding: &inferencev1alpha1.GPUShardingSpec{
-									Strategy: "layer",
-								},
-							},
-						},
-						Resources: &inferencev1alpha1.ResourceRequirements{
-							CPU:    "4",
-							Memory: "16Gi",
-						},
-					},
-				}
-				Expect(k8sClient.Create(ctx, modelResource)).To(Succeed())
-			}
-
-			By("creating a multi-GPU InferenceService")
-			isvc := &inferencev1alpha1.InferenceService{}
-			err = k8sClient.Get(ctx, serviceNamespacedName, isvc)
-			if err != nil && errors.IsNotFound(err) {
-				replicas := int32(1)
-				resource := &inferencev1alpha1.InferenceService{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      multiGPUServiceName,
-						Namespace: "default",
-					},
-					Spec: inferencev1alpha1.InferenceServiceSpec{
-						ModelRef: multiGPUModelName,
-						Replicas: &replicas,
-						Image:    "ghcr.io/ggml-org/llama.cpp:server-cuda",
-						Resources: &inferencev1alpha1.InferenceResourceRequirements{
-							GPU:       2,
-							GPUMemory: "16Gi",
-							CPU:       "4",
-							Memory:    "8Gi",
-						},
-					},
-				}
-				Expect(k8sClient.Create(ctx, resource)).To(Succeed())
-			}
-		})
-
-		AfterEach(func() {
-			By("cleaning up the multi-GPU InferenceService")
-			isvc := &inferencev1alpha1.InferenceService{}
-			err := k8sClient.Get(ctx, serviceNamespacedName, isvc)
-			if err == nil {
-				Expect(k8sClient.Delete(ctx, isvc)).To(Succeed())
-			}
-
-			By("cleaning up the multi-GPU Model")
-			model := &inferencev1alpha1.Model{}
-			err = k8sClient.Get(ctx, modelNamespacedName, model)
-			if err == nil {
-				Expect(k8sClient.Delete(ctx, model)).To(Succeed())
-			}
-
-			By("cleaning up any created Deployment")
-			deployment := &appsv1.Deployment{}
-			deploymentName := types.NamespacedName{
-				Name:      multiGPUServiceName,
-				Namespace: "default",
-			}
-			err = k8sClient.Get(ctx, deploymentName, deployment)
-			if err == nil {
-				Expect(k8sClient.Delete(ctx, deployment)).To(Succeed())
-			}
-		})
-
-		It("should create deployment with correct multi-GPU configuration", func() {
-			By("reconciling the InferenceService")
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-
-			// First reconcile may not create deployment if model isn't ready
-			// We're testing that the controller doesn't error
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: serviceNamespacedName,
-			})
-			// May return error since model download will fail (test URL)
-			// but should not panic
-			_ = err
-
-			By("verifying the InferenceService was created")
-			isvc := &inferencev1alpha1.InferenceService{}
-			err = k8sClient.Get(ctx, serviceNamespacedName, isvc)
-			Expect(err).NotTo(HaveOccurred())
-			Expect(isvc.Spec.Resources.GPU).To(Equal(int32(2)))
-		})
-	})
-})
-
-var _ = Describe("sanitizeDNSName", func() {
-	It("should replace dots with dashes", func() {
-		Expect(sanitizeDNSName("my.model.v1")).To(Equal("my-model-v1"))
-	})
-	It("should leave names without dots unchanged", func() {
-		Expect(sanitizeDNSName("my-service")).To(Equal("my-service"))
-	})
-})
 
 var _ = Describe("buildModelInitCommand", func() {
 	It("should generate cached remote download command with env var references", func() {
@@ -1558,480 +1329,6 @@ var _ = Describe("buildModelInitCommand", func() {
 	})
 })
 
-var _ = Describe("buildCachedStorageConfig", func() {
-	It("should configure PVC volume and init container for remote model", func() {
-		model := &inferencev1alpha1.Model{
-			Spec: inferencev1alpha1.ModelSpec{
-				Source: "https://example.com/model.gguf",
-			},
-			Status: inferencev1alpha1.ModelStatus{
-				CacheKey: "abc123def456",
-			},
-		}
-		config := buildCachedStorageConfig(model, "", "curl:8.18.0")
-
-		Expect(config.modelPath).To(Equal("/models/abc123def456/model.gguf"))
-		Expect(config.volumes).To(HaveLen(1))
-		Expect(config.volumes[0].Name).To(Equal("model-cache"))
-		Expect(config.volumes[0].PersistentVolumeClaim.ClaimName).To(Equal(ModelCachePVCName))
-		Expect(config.initContainers).To(HaveLen(1))
-		Expect(config.initContainers[0].Name).To(Equal("model-downloader"))
-		Expect(config.initContainers[0].Image).To(Equal("curl:8.18.0"))
-		Expect(config.volumeMounts[0].MountPath).To(Equal("/models"))
-		Expect(config.volumeMounts[0].ReadOnly).To(BeTrue())
-
-		// Verify env vars are set on the init container
-		env := config.initContainers[0].Env
-		Expect(env).To(HaveLen(3))
-		Expect(env[0]).To(Equal(corev1.EnvVar{Name: "MODEL_SOURCE", Value: "https://example.com/model.gguf"}))
-		Expect(env[1]).To(Equal(corev1.EnvVar{Name: "CACHE_DIR", Value: "/models/abc123def456"}))
-		Expect(env[2]).To(Equal(corev1.EnvVar{Name: "MODEL_PATH", Value: "/models/abc123def456/model.gguf"}))
-
-		// Verify the command does not contain the raw source URL
-		Expect(config.initContainers[0].Command[2]).NotTo(ContainSubstring("example.com"))
-	})
-
-	It("should add host-model volume for local source", func() {
-		model := &inferencev1alpha1.Model{
-			Spec: inferencev1alpha1.ModelSpec{
-				Source: "file:///mnt/models/test.gguf",
-			},
-			Status: inferencev1alpha1.ModelStatus{
-				CacheKey: "abc123",
-			},
-		}
-		config := buildCachedStorageConfig(model, "", "curl:8.18.0")
-
-		Expect(config.volumes).To(HaveLen(2))
-		Expect(config.volumes[1].Name).To(Equal("host-model"))
-		Expect(config.volumes[1].HostPath.Path).To(Equal("/mnt/models/test.gguf"))
-
-		// Verify env vars are set
-		env := config.initContainers[0].Env
-		Expect(env).To(HaveLen(3))
-		Expect(env[0]).To(Equal(corev1.EnvVar{Name: "MODEL_SOURCE", Value: "file:///mnt/models/test.gguf"}))
-	})
-
-	It("should add CA cert volume when caCertConfigMap is set", func() {
-		model := &inferencev1alpha1.Model{
-			Spec: inferencev1alpha1.ModelSpec{
-				Source: "https://example.com/model.gguf",
-			},
-			Status: inferencev1alpha1.ModelStatus{
-				CacheKey: "abc123",
-			},
-		}
-		config := buildCachedStorageConfig(model, "my-ca-certs", "curl:8.18.0")
-
-		var found bool
-		for _, v := range config.volumes {
-			if v.Name == "custom-ca-cert" {
-				found = true
-				Expect(v.ConfigMap.Name).To(Equal("my-ca-certs"))
-			}
-		}
-		Expect(found).To(BeTrue())
-		Expect(config.initContainers[0].Command[2]).To(ContainSubstring("CURL_CA_BUNDLE=/custom-certs/"))
-	})
-})
-
-var _ = Describe("buildEmptyDirStorageConfig", func() {
-	It("should configure emptyDir volume for remote model", func() {
-		model := &inferencev1alpha1.Model{
-			ObjectMeta: metav1.ObjectMeta{Name: "my-model"},
-			Spec:       inferencev1alpha1.ModelSpec{Source: "https://example.com/model.gguf"},
-		}
-		config := buildEmptyDirStorageConfig(model, "default", "", "curl:8.18.0")
-
-		Expect(config.modelPath).To(Equal("/models/default-my-model.gguf"))
-		Expect(config.volumes).To(HaveLen(1))
-		Expect(config.volumes[0].Name).To(Equal("model-storage"))
-		Expect(config.volumes[0].EmptyDir).NotTo(BeNil())
-
-		// Verify env vars are set on the init container
-		env := config.initContainers[0].Env
-		Expect(env).To(HaveLen(3))
-		Expect(env[0]).To(Equal(corev1.EnvVar{Name: "MODEL_SOURCE", Value: "https://example.com/model.gguf"}))
-		Expect(env[1]).To(Equal(corev1.EnvVar{Name: "CACHE_DIR", Value: ""}))
-		Expect(env[2]).To(Equal(corev1.EnvVar{Name: "MODEL_PATH", Value: "/models/default-my-model.gguf"}))
-
-		// Verify the command does not contain the raw source URL
-		Expect(config.initContainers[0].Command[2]).NotTo(ContainSubstring("example.com"))
-	})
-
-	It("should add CA cert volume when caCertConfigMap is set", func() {
-		model := &inferencev1alpha1.Model{
-			ObjectMeta: metav1.ObjectMeta{Name: "my-model"},
-			Spec:       inferencev1alpha1.ModelSpec{Source: "https://example.com/model.gguf"},
-		}
-		config := buildEmptyDirStorageConfig(model, "default", "my-ca-certs", "curl:8.18.0")
-
-		var found bool
-		for _, v := range config.volumes {
-			if v.Name == "custom-ca-cert" {
-				found = true
-				Expect(v.ConfigMap.Name).To(Equal("my-ca-certs"))
-			}
-		}
-		Expect(found).To(BeTrue())
-		Expect(config.initContainers[0].Command[2]).To(ContainSubstring("CURL_CA_BUNDLE=/custom-certs/"))
-	})
-})
-
-var _ = Describe("buildPVCStorageConfig", func() {
-	It("should configure PVC volume with correct claim name and path", func() {
-		model := &inferencev1alpha1.Model{
-			ObjectMeta: metav1.ObjectMeta{Name: "pvc-model"},
-			Spec:       inferencev1alpha1.ModelSpec{Source: "pvc://my-models/llama/model.gguf"},
-		}
-		config := buildPVCStorageConfig(model)
-
-		Expect(config.modelPath).To(Equal("/model-source/llama/model.gguf"))
-		Expect(config.initContainers).To(BeEmpty())
-		Expect(config.volumes).To(HaveLen(1))
-		Expect(config.volumes[0].Name).To(Equal("model-source"))
-		Expect(config.volumes[0].PersistentVolumeClaim).NotTo(BeNil())
-		Expect(config.volumes[0].PersistentVolumeClaim.ClaimName).To(Equal("my-models"))
-		Expect(config.volumes[0].PersistentVolumeClaim.ReadOnly).To(BeTrue())
-		Expect(config.volumeMounts).To(HaveLen(1))
-		Expect(config.volumeMounts[0].Name).To(Equal("model-source"))
-		Expect(config.volumeMounts[0].MountPath).To(Equal("/model-source"))
-		Expect(config.volumeMounts[0].ReadOnly).To(BeTrue())
-	})
-
-	It("should handle simple file at root of PVC", func() {
-		model := &inferencev1alpha1.Model{
-			ObjectMeta: metav1.ObjectMeta{Name: "pvc-model-simple"},
-			Spec:       inferencev1alpha1.ModelSpec{Source: "pvc://storage/model.gguf"},
-		}
-		config := buildPVCStorageConfig(model)
-
-		Expect(config.modelPath).To(Equal("/model-source/model.gguf"))
-		Expect(config.volumes[0].PersistentVolumeClaim.ClaimName).To(Equal("storage"))
-	})
-})
-
-var _ = Describe("buildModelStorageConfig PVC dispatch", func() {
-	It("should dispatch to PVC storage config when source is pvc://", func() {
-		model := &inferencev1alpha1.Model{
-			ObjectMeta: metav1.ObjectMeta{Name: "dispatch-test"},
-			Spec:       inferencev1alpha1.ModelSpec{Source: "pvc://my-claim/model.gguf"},
-			Status:     inferencev1alpha1.ModelStatus{CacheKey: "abc123"},
-		}
-		config := buildModelStorageConfig(model, "default", true, "", "curl:8.18.0")
-
-		// Should use PVC config, not cached config
-		Expect(config.volumes[0].Name).To(Equal("model-source"))
-		Expect(config.volumes[0].PersistentVolumeClaim.ClaimName).To(Equal("my-claim"))
-		Expect(config.initContainers).To(BeEmpty())
-	})
-})
-
-var _ = Describe("constructService", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{
-			Client: k8sClient,
-			Scheme: k8sClient.Scheme(),
-		}
-	})
-
-	It("should create ClusterIP service with default port", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-		}
-		svc := reconciler.constructService(isvc)
-
-		Expect(svc.Name).To(Equal("test-svc"))
-		Expect(svc.Spec.Type).To(Equal(corev1.ServiceTypeClusterIP))
-		Expect(svc.Spec.Ports[0].Port).To(Equal(int32(8080)))
-	})
-
-	It("should create NodePort service", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				Endpoint: &inferencev1alpha1.EndpointSpec{Type: "NodePort"},
-			},
-		}
-		svc := reconciler.constructService(isvc)
-		Expect(svc.Spec.Type).To(Equal(corev1.ServiceTypeNodePort))
-	})
-
-	It("should create LoadBalancer service", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				Endpoint: &inferencev1alpha1.EndpointSpec{Type: "LoadBalancer"},
-			},
-		}
-		svc := reconciler.constructService(isvc)
-		Expect(svc.Spec.Type).To(Equal(corev1.ServiceTypeLoadBalancer))
-	})
-
-	It("should use custom port", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				Endpoint: &inferencev1alpha1.EndpointSpec{Port: 3000},
-			},
-		}
-		svc := reconciler.constructService(isvc)
-		Expect(svc.Spec.Ports[0].Port).To(Equal(int32(3000)))
-	})
-
-	It("should sanitize service name with dots", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "my.model.v1", Namespace: "default"},
-		}
-		svc := reconciler.constructService(isvc)
-		Expect(svc.Name).To(Equal("my-model-v1"))
-	})
-})
-
-var _ = Describe("constructEndpoint", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{
-			Client: k8sClient,
-			Scheme: k8sClient.Scheme(),
-		}
-	})
-
-	It("should construct default endpoint URL", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-		}
-		svc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-		}
-		endpoint := reconciler.constructEndpoint(isvc, svc)
-		Expect(endpoint).To(Equal("http://test-svc.default.svc.cluster.local:8080/v1/chat/completions"))
-	})
-
-	It("should use custom port", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				Endpoint: &inferencev1alpha1.EndpointSpec{Port: 9090},
-			},
-		}
-		svc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-		}
-		endpoint := reconciler.constructEndpoint(isvc, svc)
-		Expect(endpoint).To(ContainSubstring(":9090"))
-	})
-
-	It("should use custom path", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				Endpoint: &inferencev1alpha1.EndpointSpec{Path: "/api/generate"},
-			},
-		}
-		svc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-svc", Namespace: "default"},
-		}
-		endpoint := reconciler.constructEndpoint(isvc, svc)
-		Expect(endpoint).To(HaveSuffix("/api/generate"))
-	})
-})
-
-var _ = Describe("determinePhase", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{
-			Client: k8sClient,
-			Scheme: k8sClient.Scheme(),
-		}
-	})
-
-	It("should return Ready when readyReplicas equals desiredReplicas", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test", Namespace: "default"},
-		}
-		phase, info := reconciler.determinePhase(context.Background(), isvc, 2, 2, false, &appsv1.Deployment{})
-		Expect(phase).To(Equal("Ready"))
-		Expect(info).To(BeNil())
-	})
-
-	It("should return Progressing when partially ready", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test", Namespace: "default"},
-		}
-		phase, info := reconciler.determinePhase(context.Background(), isvc, 1, 3, false, &appsv1.Deployment{})
-		Expect(phase).To(Equal("Progressing"))
-		Expect(info).To(BeNil())
-	})
-
-	It("should return Creating when no replicas ready and no scheduling issues", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "no-pods-test", Namespace: "default"},
-		}
-		phase, _ := reconciler.determinePhase(context.Background(), isvc, 0, 1, false, &appsv1.Deployment{})
-		Expect(phase).To(Equal("Creating"))
-	})
-
-	It("should return Creating when deployment is nil (Metal path)", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "test", Namespace: "default"},
-		}
-		phase, _ := reconciler.determinePhase(context.Background(), isvc, 0, 1, true, nil)
-		Expect(phase).To(Equal("Creating"))
-	})
-})
-
-var _ = Describe("resolvePriorityClassName", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
-	})
-
-	It("should return custom PriorityClassName when explicitly set", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			Spec: inferencev1alpha1.InferenceServiceSpec{PriorityClassName: "my-custom-priority"},
-		}
-		Expect(reconciler.resolvePriorityClassName(isvc)).To(Equal("my-custom-priority"))
-	})
-
-	DescribeTable("should map priority levels",
-		func(priority, expected string) {
-			isvc := &inferencev1alpha1.InferenceService{
-				Spec: inferencev1alpha1.InferenceServiceSpec{Priority: priority},
-			}
-			Expect(reconciler.resolvePriorityClassName(isvc)).To(Equal(expected))
-		},
-		Entry("critical", "critical", "llmkube-critical"),
-		Entry("high", "high", "llmkube-high"),
-		Entry("normal", "normal", "llmkube-normal"),
-		Entry("low", "low", "llmkube-low"),
-		Entry("batch", "batch", "llmkube-batch"),
-		Entry("empty defaults to normal", "", "llmkube-normal"),
-		Entry("unknown defaults to normal", "unknown", "llmkube-normal"),
-	)
-})
-
-var _ = Describe("resolveEffectivePriority", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
-	})
-
-	DescribeTable("should resolve priority values",
-		func(priority string, expected int32) {
-			isvc := &inferencev1alpha1.InferenceService{
-				Spec: inferencev1alpha1.InferenceServiceSpec{Priority: priority},
-			}
-			Expect(reconciler.resolveEffectivePriority(isvc)).To(Equal(expected))
-		},
-		Entry("critical", "critical", int32(1000000)),
-		Entry("high", "high", int32(100000)),
-		Entry("normal", "normal", int32(10000)),
-		Entry("low", "low", int32(1000)),
-		Entry("batch", "batch", int32(100)),
-		Entry("empty defaults to normal", "", int32(10000)),
-		Entry("unknown defaults to normal", "unknown", int32(10000)),
-	)
-})
-
-func deletePVCForcibly(ctx context.Context, namespace string) {
-	pvc := &corev1.PersistentVolumeClaim{}
-	pvcKey := types.NamespacedName{Name: ModelCachePVCName, Namespace: namespace}
-	if err := k8sClient.Get(ctx, pvcKey, pvc); err != nil {
-		return
-	}
-	if len(pvc.Finalizers) > 0 {
-		pvc.Finalizers = nil
-		_ = k8sClient.Update(ctx, pvc)
-	}
-	_ = k8sClient.Delete(ctx, pvc)
-	Eventually(func() bool {
-		return errors.IsNotFound(k8sClient.Get(ctx, pvcKey, &corev1.PersistentVolumeClaim{}))
-	}, "5s", "100ms").Should(BeTrue())
-}
-
-var _ = Describe("ensureModelCachePVC", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		deletePVCForcibly(context.Background(), "default")
-		reconciler = &InferenceServiceReconciler{
-			Client: k8sClient,
-			Scheme: k8sClient.Scheme(),
-		}
-	})
-
-	AfterEach(func() {
-		deletePVCForcibly(context.Background(), "default")
-	})
-
-	It("should create PVC with default 100Gi and ReadWriteOnce", func() {
-		err := reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).NotTo(HaveOccurred())
-
-		pvc := &corev1.PersistentVolumeClaim{}
-		err = k8sClient.Get(context.Background(), types.NamespacedName{Name: ModelCachePVCName, Namespace: "default"}, pvc)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(pvc.Spec.AccessModes).To(ContainElement(corev1.ReadWriteOnce))
-		storageReq := pvc.Spec.Resources.Requests[corev1.ResourceStorage]
-		Expect(storageReq.String()).To(Equal("100Gi"))
-		Expect(pvc.Labels["app.kubernetes.io/name"]).To(Equal("llmkube"))
-	})
-
-	It("should create PVC with custom size", func() {
-		reconciler.ModelCacheSize = "50Gi"
-		err := reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).NotTo(HaveOccurred())
-
-		pvc := &corev1.PersistentVolumeClaim{}
-		err = k8sClient.Get(context.Background(), types.NamespacedName{Name: ModelCachePVCName, Namespace: "default"}, pvc)
-		Expect(err).NotTo(HaveOccurred())
-		storageReq := pvc.Spec.Resources.Requests[corev1.ResourceStorage]
-		Expect(storageReq.String()).To(Equal("50Gi"))
-	})
-
-	It("should create PVC with ReadWriteMany when configured", func() {
-		reconciler.ModelCacheAccessMode = "ReadWriteMany"
-		err := reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).NotTo(HaveOccurred())
-
-		pvc := &corev1.PersistentVolumeClaim{}
-		err = k8sClient.Get(context.Background(), types.NamespacedName{Name: ModelCachePVCName, Namespace: "default"}, pvc)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(pvc.Spec.AccessModes).To(ContainElement(corev1.ReadWriteMany))
-	})
-
-	It("should set StorageClassName when configured", func() {
-		reconciler.ModelCacheClass = "fast-ssd"
-		err := reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).NotTo(HaveOccurred())
-
-		pvc := &corev1.PersistentVolumeClaim{}
-		err = k8sClient.Get(context.Background(), types.NamespacedName{Name: ModelCachePVCName, Namespace: "default"}, pvc)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(*pvc.Spec.StorageClassName).To(Equal("fast-ssd"))
-	})
-
-	It("should not error if PVC already exists", func() {
-		err := reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).NotTo(HaveOccurred())
-		err = reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).NotTo(HaveOccurred())
-	})
-
-	It("should return error for invalid cache size", func() {
-		reconciler.ModelCacheSize = "not-a-size"
-		err := reconciler.ensureModelCachePVC(context.Background(), "default")
-		Expect(err).To(HaveOccurred())
-		Expect(err.Error()).To(ContainSubstring("invalid cache size"))
-	})
-})
 
 var _ = Describe("constructDeployment additional cases", func() {
 	var reconciler *InferenceServiceReconciler
@@ -2162,646 +1459,6 @@ var _ = Describe("constructDeployment additional cases", func() {
 	})
 })
 
-var _ = Describe("findInferenceServiceForPod", func() {
-	It("should return reconcile request when pod has service label", func() {
-		reconciler := &InferenceServiceReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "test-pod",
-				Namespace: "default",
-				Labels: map[string]string{
-					"inference.llmkube.dev/service": "my-service",
-				},
-			},
-		}
-		requests := reconciler.findInferenceServiceForPod(context.Background(), pod)
-		Expect(requests).To(HaveLen(1))
-		Expect(requests[0].Name).To(Equal("my-service"))
-		Expect(requests[0].Namespace).To(Equal("default"))
-	})
-
-	It("should return nil when pod lacks service label", func() {
-		reconciler := &InferenceServiceReconciler{Client: k8sClient, Scheme: k8sClient.Scheme()}
-		pod := &corev1.Pod{
-			ObjectMeta: metav1.ObjectMeta{Name: "test-pod", Namespace: "default"},
-		}
-		requests := reconciler.findInferenceServiceForPod(context.Background(), pod)
-		Expect(requests).To(BeNil())
-	})
-})
-
-var _ = Describe("Reconcile lifecycle", func() {
-	It("should return empty result when InferenceService is not found", func() {
-		reconciler := &InferenceServiceReconciler{
-			Client:             k8sClient,
-			Scheme:             k8sClient.Scheme(),
-			InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-		}
-		result, err := reconciler.Reconcile(context.Background(), reconcile.Request{
-			NamespacedName: types.NamespacedName{Name: "nonexistent", Namespace: "default"},
-		})
-		Expect(err).NotTo(HaveOccurred())
-		Expect(result).To(Equal(reconcile.Result{}))
-	})
-
-	Context("with envtest resources", func() {
-		ctx := context.Background()
-
-		It("should set Failed status when referenced Model does not exist", func() {
-			isvcName := "isvc-no-model"
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: "nonexistent-model",
-					Replicas: &replicas,
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Phase).To(Equal(PhaseFailed))
-		})
-
-		It("should set Pending status when Model is not Ready", func() {
-			modelName := "model-not-ready"
-			isvcName := "isvc-pending"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "cpu"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Phase).To(Equal("Pending"))
-		})
-
-		It("should create Deployment and Service when Model is Ready", func() {
-			modelName := "model-ready-deploy"
-			isvcName := "isvc-deploy"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "cpu"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-					Image:    "ghcr.io/ggml-org/llama.cpp:server",
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-				dep := &appsv1.Deployment{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, dep); err == nil {
-					_ = k8sClient.Delete(ctx, dep)
-				}
-				svc := &corev1.Service{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, svc); err == nil {
-					_ = k8sClient.Delete(ctx, svc)
-				}
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			By("verifying Deployment was created")
-			dep := &appsv1.Deployment{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, dep)).To(Succeed())
-			Expect(dep.OwnerReferences).To(HaveLen(1))
-			Expect(*dep.OwnerReferences[0].Controller).To(BeTrue())
-
-			By("verifying Service was created")
-			svc := &corev1.Service{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, svc)).To(Succeed())
-			Expect(svc.OwnerReferences).To(HaveLen(1))
-
-			By("verifying status was updated")
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Phase).To(Equal("Creating"))
-			Expect(updated.Status.Endpoint).NotTo(BeEmpty())
-		})
-
-		It("should skip Deployment for Metal accelerator", func() {
-			modelName := "metal-model"
-			isvcName := "isvc-metal"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "metal"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			By("verifying no Deployment was created")
-			dep := &appsv1.Deployment{}
-			err = k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, dep)
-			Expect(errors.IsNotFound(err)).To(BeTrue())
-
-			By("verifying no Service was created (Metal Agent manages its own)")
-			svc := &corev1.Service{}
-			err = k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, svc)
-			Expect(errors.IsNotFound(err)).To(BeTrue())
-
-			By("verifying status is Ready (Metal returns desiredReplicas as ready)")
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Phase).To(Equal("Ready"))
-		})
-
-		It("should set correct endpoint URL for Metal InferenceService", func() {
-			modelName := "metal-endpoint-model"
-			isvcName := "isvc-metal-endpoint"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "metal"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Endpoint).To(Equal(
-				"http://isvc-metal-endpoint.default.svc.cluster.local:8080/v1/chat/completions",
-			))
-		})
-
-		It("should set DNS-sanitized endpoint URL for Metal InferenceService with dots in name", func() {
-			modelName := "metal-dot-model"
-			isvcName := "isvc-metal.v1.0"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "metal"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Endpoint).To(Equal(
-				"http://isvc-metal-v1-0.default.svc.cluster.local:8080/v1/chat/completions",
-			))
-		})
-
-		It("should use custom endpoint port and path for Metal InferenceService", func() {
-			modelName := "metal-custom-ep-model"
-			isvcName := "isvc-metal-custom-ep"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "metal"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-					Endpoint: &inferencev1alpha1.EndpointSpec{
-						Port: 9090,
-						Path: "/api/generate",
-					},
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			updated := &inferencev1alpha1.InferenceService{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, updated)).To(Succeed())
-			Expect(updated.Status.Endpoint).To(Equal(
-				"http://isvc-metal-custom-ep.default.svc.cluster.local:9090/api/generate",
-			))
-		})
-
-		It("should default replicas to 1 when nil", func() {
-			modelName := "model-nil-replicas"
-			isvcName := "isvc-nil-replicas"
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "cpu"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					// Replicas intentionally nil
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-				dep := &appsv1.Deployment{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, dep); err == nil {
-					_ = k8sClient.Delete(ctx, dep)
-				}
-				svc := &corev1.Service{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, svc); err == nil {
-					_ = k8sClient.Delete(ctx, svc)
-				}
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			dep := &appsv1.Deployment{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, dep)).To(Succeed())
-			Expect(*dep.Spec.Replicas).To(Equal(int32(1)))
-		})
-
-		It("should create PVC when model has CacheKey and ModelCachePath is set", func() {
-			modelName := "model-with-cache"
-			isvcName := fmt.Sprintf("isvc-pvc-test-%d", GinkgoRandomSeed())
-
-			model := &inferencev1alpha1.Model{
-				ObjectMeta: metav1.ObjectMeta{Name: modelName, Namespace: "default"},
-				Spec: inferencev1alpha1.ModelSpec{
-					Source:   "https://example.com/model.gguf",
-					Hardware: &inferencev1alpha1.HardwareSpec{Accelerator: "cpu"},
-				},
-			}
-			Expect(k8sClient.Create(ctx, model)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, model)
-			}()
-
-			model.Status.Phase = PhaseReady
-			model.Status.CacheKey = "abc123def456"
-			Expect(k8sClient.Status().Update(ctx, model)).To(Succeed())
-
-			replicas := int32(1)
-			isvc := &inferencev1alpha1.InferenceService{
-				ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-				Spec: inferencev1alpha1.InferenceServiceSpec{
-					ModelRef: modelName,
-					Replicas: &replicas,
-				},
-			}
-			Expect(k8sClient.Create(ctx, isvc)).To(Succeed())
-			defer func() {
-				_ = k8sClient.Delete(ctx, isvc)
-				dep := &appsv1.Deployment{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, dep); err == nil {
-					_ = k8sClient.Delete(ctx, dep)
-				}
-				svc := &corev1.Service{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: isvcName, Namespace: "default"}, svc); err == nil {
-					_ = k8sClient.Delete(ctx, svc)
-				}
-				pvc := &corev1.PersistentVolumeClaim{}
-				if err := k8sClient.Get(ctx, types.NamespacedName{Name: ModelCachePVCName, Namespace: "default"}, pvc); err == nil {
-					_ = k8sClient.Delete(ctx, pvc)
-				}
-			}()
-
-			reconciler := &InferenceServiceReconciler{
-				Client:             k8sClient,
-				Scheme:             k8sClient.Scheme(),
-				InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-				ModelCachePath:     "/models",
-			}
-			_, err := reconciler.Reconcile(ctx, reconcile.Request{
-				NamespacedName: types.NamespacedName{Name: isvcName, Namespace: "default"},
-			})
-			Expect(err).NotTo(HaveOccurred())
-
-			pvc := &corev1.PersistentVolumeClaim{}
-			Expect(k8sClient.Get(ctx, types.NamespacedName{Name: ModelCachePVCName, Namespace: "default"}, pvc)).To(Succeed())
-		})
-	})
-})
-
-var _ = Describe("reconcileService Metal path", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{
-			Client:             k8sClient,
-			Scheme:             k8sClient.Scheme(),
-			InitContainerImage: "docker.io/curlimages/curl:8.18.0",
-		}
-	})
-
-	It("should return minimal Service with correct name and namespace when isMetal is true", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "metal-svc-test", Namespace: "default"},
-			Spec:       inferencev1alpha1.InferenceServiceSpec{ModelRef: "some-model"},
-		}
-
-		svc, result, err := reconciler.reconcileService(context.Background(), isvc, true, 1, true)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(result).To(BeNil())
-		Expect(svc).NotTo(BeNil())
-		Expect(svc.Name).To(Equal("metal-svc-test"))
-		Expect(svc.Namespace).To(Equal("default"))
-	})
-
-	It("should DNS-sanitize the minimal Service name when isMetal is true", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "llama-3.2-3b", Namespace: "test-ns"},
-			Spec:       inferencev1alpha1.InferenceServiceSpec{ModelRef: "some-model"},
-		}
-
-		svc, result, err := reconciler.reconcileService(context.Background(), isvc, true, 1, true)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(result).To(BeNil())
-		Expect(svc.Name).To(Equal("llama-3-2-3b"))
-		Expect(svc.Namespace).To(Equal("test-ns"))
-	})
-
-	It("should not have Spec fields populated on the minimal Service", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "metal-minimal", Namespace: "default"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				ModelRef: "some-model",
-				Endpoint: &inferencev1alpha1.EndpointSpec{Port: 9090, Type: "LoadBalancer"},
-			},
-		}
-
-		svc, _, err := reconciler.reconcileService(context.Background(), isvc, true, 1, true)
-		Expect(err).NotTo(HaveOccurred())
-		Expect(svc.Spec.Ports).To(BeEmpty())
-		Expect(svc.Spec.Type).To(Equal(corev1.ServiceType("")))
-	})
-
-	It("should not create any K8s Service resource when isMetal is true", func() {
-		isvcName := "metal-no-k8s-svc"
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: isvcName, Namespace: "default"},
-			Spec:       inferencev1alpha1.InferenceServiceSpec{ModelRef: "some-model"},
-		}
-
-		_, _, err := reconciler.reconcileService(context.Background(), isvc, true, 1, true)
-		Expect(err).NotTo(HaveOccurred())
-
-		svc := &corev1.Service{}
-		err = k8sClient.Get(context.Background(), types.NamespacedName{Name: isvcName, Namespace: "default"}, svc)
-		Expect(errors.IsNotFound(err)).To(BeTrue())
-	})
-})
-
-var _ = Describe("constructEndpoint with Metal minimal Service", func() {
-	var reconciler *InferenceServiceReconciler
-
-	BeforeEach(func() {
-		reconciler = &InferenceServiceReconciler{
-			Client: k8sClient,
-			Scheme: k8sClient.Scheme(),
-		}
-	})
-
-	It("should construct correct URL from minimal Metal Service with default settings", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "metal-test", Namespace: "default"},
-		}
-		minimalSvc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      sanitizeDNSName(isvc.Name),
-				Namespace: isvc.Namespace,
-			},
-		}
-		endpoint := reconciler.constructEndpoint(isvc, minimalSvc)
-		Expect(endpoint).To(Equal("http://metal-test.default.svc.cluster.local:8080/v1/chat/completions"))
-	})
-
-	It("should construct correct URL from minimal Metal Service with custom port and path", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "metal-custom", Namespace: "production"},
-			Spec: inferencev1alpha1.InferenceServiceSpec{
-				Endpoint: &inferencev1alpha1.EndpointSpec{
-					Port: 3000,
-					Path: "/api/v2/infer",
-				},
-			},
-		}
-		minimalSvc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      sanitizeDNSName(isvc.Name),
-				Namespace: isvc.Namespace,
-			},
-		}
-		endpoint := reconciler.constructEndpoint(isvc, minimalSvc)
-		Expect(endpoint).To(Equal("http://metal-custom.production.svc.cluster.local:3000/api/v2/infer"))
-	})
-
-	It("should construct correct URL when Metal Service name is DNS-sanitized", func() {
-		isvc := &inferencev1alpha1.InferenceService{
-			ObjectMeta: metav1.ObjectMeta{Name: "model.v2.1", Namespace: "ml"},
-		}
-		minimalSvc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      sanitizeDNSName(isvc.Name),
-				Namespace: isvc.Namespace,
-			},
-		}
-		endpoint := reconciler.constructEndpoint(isvc, minimalSvc)
-		Expect(endpoint).To(Equal("http://model-v2-1.ml.svc.cluster.local:8080/v1/chat/completions"))
-	})
-})
 
 var _ = Describe("Security Context Configuration", func() {
 	var (
@@ -2978,5 +1635,5 @@ var _ = Describe("Security Context Configuration", func() {
 	})
 })
 
-// boolPtr64 is a test helper for creating *int64 inline
+
 func boolPtr64(v int64) *int64 { return &v }
